Restructure gulp folder; remove null's script thing
