--- tgui/Gulpfile.js
@@ -1,2 +1,0 @@
-require('babel-register')
-require('./gulp')
