--- tgui/gulp/.babelrc
@@ -1,5 +1,0 @@
-{
-  "presets": [
-    "es2015"
-  ]
-}
