+++ tgui/src/.babelrc
@@ -1,0 +1,8 @@
+{
+  "presets": [
+    "es2015-loose"
+  ],
+  "plugins": [
+    "external-helpers"
+  ]
+}
