+++ tgui/src/styles/nanotrasen.styl
@@ -1,5 +1,5 @@
 body.nanotrasen
-  background: data-url('nanotrasen.svg') no-repeat fixed center/70% 70%,
+  background: data-url('images/nanotrasen.svg') no-repeat fixed center/70% 70%,
     linear-gradient(to bottom,
       background-color-start 0%,
       background-color-end 100%)
