+++ tgui/src/styles/syndicate.styl
@@ -19,7 +19,7 @@ body.syndicate
   button-color-danger      = yellow
 
 
-  background: data-url('syndicate.svg') no-repeat fixed center/70% 70%,
+  background: data-url('images/syndicate.svg') no-repeat fixed center/70% 70%,
     linear-gradient(to bottom,
       background-color-start 0%,
       background-color-end 100%)
